fix disconnections && clean remote goal functions

+++ frontend/js/pages/game/src/World.js
@@ -14,15 +14,14 @@ import { Opponent } from './components/Opponent.js';
 import { Score } from './components/Score.js';
 import { airHockeyTable } from './systems/Loader.js';
 
-import interactiveSocket from '../../home/socket.js';
 import {
-	BoxGeometry,
 	CapsuleGeometry,
 	MeshStandardMaterial,
 	SphereGeometry,
 	Vector2,
 	Vector3
 } from 'three';
+
 import { EffectComposer } from '/node_modules/three/examples/jsm/postprocessing/EffectComposer.js';
 import { RenderPass } from '/node_modules/three/examples/jsm/postprocessing/RenderPass.js';
 import { BloomPass } from '/node_modules/three/examples/jsm/postprocessing/BloomPass.js';
@@ -31,7 +30,6 @@ import { FilmPass } from '/node_modules/three/examples/jsm/postprocessing/FilmPa
 import { OutputPass } from '/node_modules/three/examples/jsm/postprocessing/OutputPass.js';
 import { GlitchPass } from '/node_modules/three/examples/jsm/postprocessing/GlitchPass.js';
 import { RenderPixelatedPass } from '/node_modules/three/examples/jsm/postprocessing/RenderPixelatedPass.js';
-import { ParticleSystem } from './components/ParticleSystem.js';
 
 const GameState = {
 	InMenu: "inMenu",
@@ -53,43 +51,24 @@ class World {
 
 		this.createContainer( container );
 		this.createGame();
-
-		/// DEBUG TEMP
-		document.addEventListener('keydown', (event) => {
-			// if ( event.code == "KeyR" ) {
-			// 	console.warn("-- DELETION! --");
-			// 	World._instance.deleteGame();
-			// }
-			// if ( event.code == "KeyE" ) {
-			// 	console.warn("-- RESUME! --");
-			// 	World._instance.createGame();
-			// }
-			// if ( event.code == "Space" && this.currentGameState == GameState.InMenu ) {
-			// 	if ( !interactiveSocket.interactive_socket )
-			// 		return console.error("interactiveSocket not up");
-			// 	console.log("-- Waiting for Opponent --");
-			// 	this.currentGameState = GameState.LookingForPlayer;
-			// 	interactiveSocket.sendMessageSocket(JSON.stringify({"type": "Find Match"}));
-			// }
-		}, false);
 	}
 
 	joinMatch( wsPath, side ) {
 		// HOST
 		if ( side == "A" ) {
-			console.log("-- Becoming Host --");
+			// console.log("-- Becoming Host --");
 			this.createSocket( '/' + wsPath, 7.2, this.terrain.rightGoalZone, this.terrain.leftGoalZone );
 		}
 		// CLIENT
 		else if ( side == "B" ) {
-			console.log("-- Becoming Client --");
+			// console.log("-- Becoming Client --");
 			this.createSocket( '/' + wsPath, -7.2, this.terrain.leftGoalZone, this.terrain.rightGoalZone );
 		}
 
 	}
 
 	createSocket( path, xpos, goalZone, opponentGoalZone ) {
-		console.log("-- Socket Created! --");
+		// console.log("-- Socket Created! --");
 		this.currentGameState = GameState.Connecting;
 
 		this.socket = new WebSocket('wss://' + window.location.host + path);
@@ -114,8 +93,8 @@ class World {
 		this.renderer = createRenderer();
 
 		this.composer = new EffectComposer( this.renderer);
-		// this.composer.addPass( new RenderPass( this.scene, this.camera ) );
-		this.composer.addPass( new RenderPixelatedPass( 2, this.scene, this.camera ) );
+		this.composer.addPass( new RenderPass( this.scene, this.camera ) );
+		// this.composer.addPass( new RenderPixelatedPass( 2, this.scene, this.camera ) );
 		// this.composer.addPass( new BloomPass( 1, 25, 4, 256 ) );
 		this.composer.addPass( new UnrealBloomPass( 256, 0.5, 0.5, 0.5 ) );
 		this.composer.addPass( new FilmPass( 0.15, false ) );
@@ -155,24 +134,6 @@ class World {
 		
 		this.camera.viewLarge( 0 );
 		this.loop.start();
-
-		const bg = new BoxGeometry( 2, 2, 2 );
-		const blue = new MeshStandardMaterial({ color: 'blue' });
-		const parameters = {
-			duration: 0.8,
-			position: new Vector3( 0, 0, 0 ),
-			positionRandomRange: new Vector3( 2, 2, 2 ),
-			direction: new Vector3( 0, 0, 0 ),
-			directionRandomRange: new Vector3( 0, 0, 0 ),
-			speed: 100,
-			speedOverTime: 0,
-			eulerRotation: new Vector3( 0, 0, 0 ),
-			eulerRotationRandomRange: new Vector3( Math.PI, Math.PI, Math.PI ),
-			scale: new Vector3( 1, 1, 1 ),
-			scaleRandomRange: new Vector3( 0, 0, 0 ),
-			scaleOverTime: new Vector3( 0, 0, 0 )
-		}
-		this.particles = new ParticleSystem( bg, blue, 1000, parameters );
 	}
 
 	initMatch() {
@@ -183,24 +144,42 @@ class World {
 		this.balls.renderer.setEnabled(true);
 		this.balls.updatable.setEnabled(true);
 		this.camera.viewTable( 1, function() {
-			this.balls.init();
+			if ( this.currentGameState === GameState.InMatch )
+				this.balls.init();
 		}.bind( this ) );
 
 		this.socket.addEventListener("message", (event) => {
+			if ( event.data === "Closing" ) {
+				console.log("Opponent Ragequited");
+				this.endMatch();
+				return;
+			}
 			const msg = JSON.parse( event.data );
-			if ( this.opponent != undefined )
+			if ( this.opponent != undefined && msg.pos != undefined )
 				this.opponent.position.copy( msg.pos );
 			if ( msg.ballInst != undefined ) {
 				if ( msg.scored == true ) {
-					this.score.increment( msg.goalScoredId );
-					this.balls.initInst( this.balls.ballInst[ msg.ballInst.id ] );
+					if ( this.terrain.leftGoalZone.paddle.isOpponent )
+						this.terrain.leftGoalZone.goal( msg.ballInst );
+					if ( this.terrain.rightGoalZone.paddle.isOpponent )
+						this.terrain.rightGoalZone.goal( msg.ballInst );
 				}
 				else
 					this.balls.overwriteInst( msg.ballInst );
 			}
 		})
 
-		console.log("-- Starting Match --");
+		// TBD via backend
+		window.addEventListener("beforeunload", (event) => {
+			this.socket.send("Closing");
+			this.endMatch();
+		})
+		window.addEventListener("popstate", (event) => {
+			this.socket.send("Closing");
+			this.endMatch();
+		})
+
+		// console.log("-- Starting Match --");
 		this.currentGameState = GameState.InMatch;
 	}
 
@@ -208,9 +187,11 @@ class World {
 		this.camera.viewLarge( 1 , function() {
 			document.getElementById('ui').classList.remove("d-none");
 		} );
-		this.player.delete();
+		if ( this.player != undefined )
+			this.player.delete();
 		this.player = undefined;
-		this.opponent.delete();
+		if ( this.opponent != undefined )
+			this.opponent.delete();
 		this.opponent = undefined;
 		this.balls.updatable.setEnabled(false);
 		this.balls.renderer.setEnabled(false);
